+++ src/ChatContext.tsx
@@ -1,5 +1,5 @@
-import { createContext, useCallback, useContext, useEffect, useRef, useState } from "react";
-import { ChatContextType, ChatProviderType, GroupChatEventType } from "./types";
+import { createContext, use, useContext, useEffect, useRef, useState } from "react";
+import { ChatContextType, ChatProviderType, ChatStatus, GroupChatEventType, MessageStatus } from "./types";
 import { Chat, ChatEvent, ChatType, Dialogs, DialogType, Messages } from "connectycube/types";
 
 import ConnectyCube from "connectycube";
@@ -25,6 +25,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
   // state
   const [isOnline, setIsOnline] = useState<boolean>(navigator.onLine);
   const [isConnected, setIsConnected] = useState(false);
+  const [chatStatus, setChatStatus] = useState(ChatStatus.DISCONNECTED);
   const [unreadMessagesCount, setUnreadMessagesCount] = useState<ChatContextType["unreadMessagesCount"]>({ total: 0 });
   const [messages, setMessages] = useState<{ [dialogId: string]: Messages.Message[] }>({});
   const [typingStatus, setTypingStatus] = useState<{ [dialogId: string]: number[] }>({});
@@ -33,6 +34,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
   const typingTimers = useRef<{ [dialogId: string]: { [userId: number | string]: NodeJS.Timeout } }>({});
   const onMessageRef = useRef<Chat.OnMessageListener | null>(null);
   const onSignalRef = useRef<Chat.OnMessageSystemListener | null>(null);
+  const onMessageSentRef = useRef<Chat.OnMessageSentListener | null>(null);
   const onMessageErrorRef = useRef<Chat.OnMessageErrorListener | null>(null);
   const privateDialogsIdsRef = useRef<{ [userId: number | string]: string }>({});
   // state refs
@@ -45,6 +47,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
   const { _retrieveAndStoreUsers } = chatUsers;
 
   const connect = async (credentials: Chat.ConnectionParams) => {
+    setChatStatus(ChatStatus.CONNECTING);
     try {
       const _isConnected = await ConnectyCube.chat.connect(credentials);
       if (_isConnected) {
@@ -52,6 +55,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
         setCurrentUserId(credentials.userId);
       }
     } catch (error) {
+      setChatStatus(ChatStatus.DISCONNECTED);
       console.error(`Failed to connect due to ${error}`);
     }
   };
@@ -156,7 +160,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
             ...attachment,
             url: ConnectyCube.storage.privateUrl(attachment.uid),
           }));
-          return { ...msg, attachments };
+          return { ...msg, attachments, status: msg.read ? MessageStatus.READ : MessageStatus.SENT };
         });
       setMessages((prevMessages) => ({ ...prevMessages, [dialogId]: retrievedMessages }));
       return retrievedMessages;
@@ -375,7 +379,9 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
     setMessages((prevMessages) => ({
       ...prevMessages,
       [dialog._id]: prevMessages[dialog._id].map((msg) =>
-        msg._id === tempId ? { ...msg, _id: messageId, attachments, isLoading: false } : msg,
+        msg._id === tempId
+          ? { ...msg, _id: messageId, attachments, isLoading: false, status: MessageStatus.WAIT }
+          : msg,
       ),
     }));
   };
@@ -456,6 +462,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
           attachments: attachments ? attachments : [],
           reactions: {} as any,
           isLoading,
+          status: MessageStatus.WAIT,
         },
       ],
     }));
@@ -471,7 +478,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
     setMessages((prevMessages) => ({
       ...prevMessages,
       [dialogId]: prevMessages[dialogId].map((message) =>
-        message._id === messageId ? { ...message, read: 1 } : message,
+        message._id === messageId ? { ...message, read: 1, status: MessageStatus.READ } : message,
       ),
     }));
 
@@ -586,6 +593,10 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
     onMessageErrorRef.current = callbackFn;
   };
 
+  const processOnMessageSent = (callbackFn: Chat.OnMessageSentListener | null) => {
+    onMessageSentRef.current = callbackFn;
+  };
+
   // Internet listeners
   useEffect(() => {
     const abortController1 = new AbortController();
@@ -617,11 +628,20 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
     };
   }, []);
 
+  useEffect(() => {
+    if (isConnected) {
+      setChatStatus(isOnline ? ChatStatus.CONNECTED : ChatStatus.CONNECTING);
+    }
+  }, [isOnline, isConnected]);
+
   const _processDisconnect = () => {
+    setIsConnected(false);
+    setChatStatus(ChatStatus.DISCONNECTED);
     setActivatedDialogs({});
   };
 
   const _processReconnect = () => {
+    setIsConnected(true);
     console.log("[useChat] Reconnected");
   };
 
@@ -675,6 +695,25 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
   const _processErrorMessage = (messageId: string, error: { code: number; info: string }) => {
     if (onMessageErrorRef.current) {
       onMessageErrorRef.current(messageId, error);
+    }
+  };
+
+  const _processSentMessage = (lost: Chat.MessageParams | null, sent?: Chat.MessageParams) => {
+    if (onMessageSentRef.current) {
+      onMessageSentRef.current(lost, sent);
+    }
+
+    const dialogId = lost?.extension.dialog_id || sent?.extension.dialog_id;
+    const messageId = lost?.id || sent?.id;
+    const messageStatus = lost ? MessageStatus.LOST : sent ? MessageStatus.SENT : MessageStatus.WAIT;
+
+    if (dialogId && messageId) {
+      setMessages((prevMessages) => ({
+        ...prevMessages,
+        [dialogId]: prevMessages[dialogId].map((message) =>
+          message._id === messageId ? { ...message, status: messageStatus } : message,
+        ),
+      }));
     }
   };
 
@@ -756,15 +795,19 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
       return;
     }
 
-    setMessages((prevMessages) => {
-      (prevMessages[dialogId] || []).forEach((message) => {
-        if (message._id === messageId && message.read === 0) {
-          message.read = 1;
-          message.read_ids?.push(userId);
-        }
-      });
-      return prevMessages;
-    });
+    setMessages((prevMessages) => ({
+      ...prevMessages,
+      [dialogId]: prevMessages[dialogId].map((message) =>
+        message._id === messageId
+          ? {
+              ...message,
+              status: MessageStatus.READ,
+              read_ids: message.read_ids ? [...new Set([...message.read_ids, userId])] : [userId],
+              read: 1,
+            }
+          : message,
+      ),
+    }));
   };
 
   const _processTypingMessageStatus = (isTyping: boolean, userId: number, dialogId: string | null) => {
@@ -801,6 +844,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
     ConnectyCube.chat.addListener(ChatEvent.RECONNECTED, _processReconnect);
     ConnectyCube.chat.addListener(ChatEvent.MESSAGE, _processMessage);
     ConnectyCube.chat.addListener(ChatEvent.ERROR_MESSAGE, _processErrorMessage);
+    ConnectyCube.chat.addListener(ChatEvent.SENT_MESSAGE, _processSentMessage);
     ConnectyCube.chat.addListener(ChatEvent.SYSTEM_MESSAGE, _processSystemMessage);
     ConnectyCube.chat.addListener(ChatEvent.READ_MESSAGE, _processReadMessageStatus);
     ConnectyCube.chat.addListener(ChatEvent.TYPING_MESSAGE, _processTypingMessageStatus);
@@ -818,8 +862,9 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
     <ChatContext.Provider
       value={{
         isOnline,
-        connect,
         isConnected,
+        chatStatus,
+        connect,
         disconnect,
         currentUserId,
         selectDialog,
@@ -847,6 +892,7 @@ export const ChatProvider = ({ children }: ChatProviderType): React.ReactElement
         processOnSignal,
         processOnMessage,
         processOnMessageError,
+        processOnMessageSent,
         ...chatBlockList,
         ...chatUsers.exports,
       }}