+++ src/types/index.ts
@@ -9,8 +9,9 @@ export interface ChatProviderType {
 
 export interface ChatContextType extends BlockListHook, UsersHookExports {
   isOnline: boolean;
-  connect: (credentials: Chat.ConnectionParams) => Promise<void>;
   isConnected: boolean;
+  chatStatus: ChatStatus;
+  connect: (credentials: Chat.ConnectionParams) => Promise<void>;
   disconnect: () => void;
   currentUserId?: number;
   createChat: (userId: number, extensions?: { [key: string]: any }) => Promise<Dialogs.Dialog>;
@@ -43,6 +44,7 @@ export interface ChatContextType extends BlockListHook, UsersHookExports {
   processOnSignal: (fn: Chat.OnMessageSystemListener | null) => void;
   processOnMessage: (fn: Chat.OnMessageListener | null) => void;
   processOnMessageError: (fn: Chat.OnMessageErrorListener | null) => void;
+  processOnMessageSent: (fn: Chat.OnMessageSentListener | null) => void;
 }
 
 export enum GroupChatEventType {
@@ -51,4 +53,17 @@ export enum GroupChatEventType {
   ADD_PARTICIPANTS = "dialog/ADD_PARTICIPANTS",
   REMOVE_PARTICIPANTS = "dialog/REMOVE_PARTICIPANTS",
   NEW_DIALOG = "dialog/NEW_DIALOG",
+}
+
+export enum MessageStatus {
+  WAIT = "wait",
+  LOST = "lost",
+  SENT = "sent",
+  READ = "read",
+}
+
+export enum ChatStatus {
+  DISCONNECTED = "disconnected",
+  CONNECTING = "connecting",
+  CONNECTED = "connected",
 }